Semantic checks to see if a DO variable is modified
I added infrastructure to SemanticsContext to track active DO variables and
the source locations where they appear in DO statements.  I also added code
to tools.[h,cc] to check to see if a DO variable is already defined, and, if
so, to emit an error message along with a reference to the relevant DO
construct.  I also added calls to several places where variables are defined
to determine if the definitions are happening in the context of an active DO
construct.